Adds PHP 8.1 support (#142)
* Adds PHP 8.1 support

* Fixes float to int

* Apply fixes from StyleCI

+++ src/Aws/AwsStorageProvider.php
@@ -38,7 +38,7 @@ class AwsStorageProvider
     {
         $stream = fopen($file, 'r+');
 
-        $size = round(filesize($file) / 1024, 2);
+        $size = (int) round(filesize($file) / 1024, 2);
 
         if ($withProgress) {
             $progressBar = new ProgressBar(Helpers::app('output'), $size);
@@ -49,7 +49,7 @@ class AwsStorageProvider
         }
 
         $progressCallback = $withProgress ? function ($_, $__, $___, $uploaded) use ($progressBar) {
-            $progressBar->setProgress(round($uploaded / 1024, 2));
+            $progressBar->setProgress((int) round($uploaded / 1024, 2));
         }
         : null;
